Change method of setting current location
Since the locations have been moved to the user object we can
now set the current location based on that object.

For now, we are only supporting one location so if there are more
than one it will pass through this middleware.

In a future change we will redirect to a locations feature where
a user can choose which location to view when they first sign in
and this value is not set.
